move optimize-related charts to charts, fix typing conflicts
